requirements.txt
flask
python-dotenv
requests
numpy
aiohttp
waitress
stripe
